+++ src/Tickets/AirplaneTicket.java
@@ -5,7 +5,7 @@ import Person.Person;
 import java.util.ArrayList;
 
 public class AirplaneTicket extends Ticket{
-    public AirplaneTicket(float price, TicketDecorator ticketDecorator, Person person) {
+    public AirplaneTicket(double price, TicketDecorator ticketDecorator, Person person) {
         super(price, ticketDecorator, person);
     }
 }
