+++ src/Tickets/DefaultTicket.java
@@ -5,7 +5,7 @@ import Person.Person;
 import java.util.ArrayList;
 
 public class DefaultTicket extends Ticket{
-    public DefaultTicket(float price, TicketDecorator ticketDecorator, Person person) {
+    public DefaultTicket(double price, TicketDecorator ticketDecorator, Person person) {
         super(price, ticketDecorator, person);
     }
 }
