+++ src/Tickets/EvenlySplitDecorator.java
@@ -7,20 +7,20 @@ import java.util.Map;
 import Person.*;
 
 public class EvenlySplitDecorator implements TicketDecorator{
-    private float price;
+    private double price;
     private Person person;
     private ArrayList<Person> paidFor;
 
-    public EvenlySplitDecorator(Person person, float price, ArrayList<Person> paidFor) {
+    public EvenlySplitDecorator(Person person, double price, ArrayList<Person> paidFor) {
         this.price = price;
         this.person = person;
         this.paidFor = paidFor;
     }
 
     @Override
-    public Map<Person,Float> getBalanceDifferencePerPerson() {
-        Map<Person,Float> differencePerPerson = new HashMap<Person, Float>();
-        float pricePerPerson = price/paidFor.size();
+    public Map<Person,Double> getBalanceDifferencePerPerson() {
+        Map<Person,Double> differencePerPerson = new HashMap<Person, Double>();
+        double pricePerPerson = price/paidFor.size();
         for (Person pers : paidFor) {
             if (!(pers == person)) {
                 differencePerPerson.put(pers, - pricePerPerson);
